tests/wpt/tests/css/css-fonts/parsing/font-synthesis-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Fonts Module: getComputedStyle().fontSynthesis</title>
<link rel="help" href="https://drafts.csswg.org/css-fonts/#font-synthesis">
<meta name="assert" content="font-synthesis computed value is as specified.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
</head>
<body>
<div id="target"></div>
<script>
test_computed_value('font-synthesis', 'none');
test_computed_value('font-synthesis', 'weight');
test_computed_value('font-synthesis', 'style');
test_computed_value('font-synthesis', 'small-caps');
test_computed_value('font-synthesis', 'position');
test_computed_value('font-synthesis', 'small-caps position');
test_computed_value('font-synthesis', 'style small-caps');
test_computed_value('font-synthesis', 'style position');
test_computed_value('font-synthesis', 'style small-caps position');
test_computed_value('font-synthesis', 'weight small-caps');
test_computed_value('font-synthesis', 'weight style');
test_computed_value('font-synthesis', 'weight position');
test_computed_value('font-synthesis', 'weight style small-caps');
test_computed_value('font-synthesis', 'weight style small-caps position');
</script>
</body>
</html>
